fix notifications with value on wayland
I didn't need to update zbus, but it was something I tried while trying
to fix this problem. Turns out, I wasn't very familiar with the XDG
notifications specification about the "value" hint, and I was sending an
unsigned 32 rather than a signed 32.

+++ src/bar_items/pulse/mod.rs
@@ -715,7 +715,7 @@ impl BarItem for Pulse {
                     }
                     Command::NotifyVolume { name, volume, mute } => {
                         if self.notify.should_notify(NotificationSetting::VolumeMute) {
-                            let _ = notifications.pulse_volume_mute(name, volume, mute).await;
+                            let _ = notifications.pulse_volume_mute(name, volume as i32, mute).await;
                         }
                     }
                     Command::NotifyNewSourceSink { name, what } => {
